requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
multidict==6.1.0
packaging==24.2
pillow==11.0.0
propcache==0.2.0
psycopg2==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.6
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
